Split lowered-program test helpers into focused modules
